feat!: add `scopes` field to `KeyringAccount` (#134)
The `scopes` field is now required on the `KeyringAccount`. This field
will be used to identify the supported scopes (using CAIP-2 chain IDs)
for a given account on MetaMask clients.

This is a breaking change since the new field is mandatory and the Snap
keyring bridge will enforce this field is present (once updated on both
clients).