perf: Speed up Solana key derivation (#191)
Improves the speed of the Solana key derivation in multiple ways:
1. Use `snap_getBip32Entropy` to derive the full key to prevent having
to do additional key derivation steps inside the Snap. This matters
because the derivation logic gets faster after initializing (during the
first rounds).
2. Use public key already derived public key from `snap_getBip32Entropy`
instead of recomputing the public key using the Solana library. This
lets us skip more cryptographic operations that we are already doing as
part of the API call above.