feat: forward `MetaMaskOptions` in `keyring_createAccount` method (#266)
We have added a new "internal context" during `keyring_createAccount`
that allows use to customize the account creation flow within the Snap
keyring.

In order to "correlate" the initial call to a `keyring_createAccount`
and its associated `notify:accountCreated` event, a new `metamask`
object must be forwarded to the event.