chore: improve error message for account creation (#286)
Instead of throwing a generic `"Error creating account"` we should throw
the real error when creating an account. This issue is that when the
Solana snap breaks in the codebase, we have no idea why and then we must
run it locally to debug. If we just threw the error directly debugging
would be much easier.

---------

Co-authored-by: Antonio Regadas <antonio.regadas@consensys.net>