chore: adds check and throws error (#471)
Two different topics in this PR:

- Checks that the account address in the request parameters matches the
account used for signing and If it doesn't match, throw the same error
MM throws when the account is not authorized. (Original solo goal)

- Fixes the `buildUrl` function, where we now add an optional parameter
called `encodePathParams` so that Price API calls can set it to `false`,
since it doesn't support the decoding of URL params.