Fix self transfer parsing (#504)
This PR addresses [an
issue](https://github.com/MetaMask/metamask-extension/issues/35431)
where self-transfers using Token2022 were incorrectly parsed.

We fix it by standardizing the way we parse instructions. Before, we
were using ad-hoc (and incomplete, hence the bug) decoding. Now we defer
the parsing to the universal instruction parser that was already in use
in other parts of the snap.